Rename dashboard/presentation to views
