+++ lib/src/features/dashboard/views/dashboard_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:librarian_app/src/features/dashboard/presentation/desktop_layout.dart';
-import 'package:librarian_app/src/features/dashboard/presentation/mobile_layout.dart';
+import 'package:librarian_app/src/features/dashboard/views/desktop_layout.dart';
+import 'package:librarian_app/src/features/dashboard/views/mobile_layout.dart';
 
 class DashboardPage extends StatelessWidget {
   const DashboardPage({super.key});
