+++ lib/src/features/splash/presentation/splash_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:librarian_app/src/features/authentication/presentation/signin_page.dart';
-import 'package:librarian_app/src/features/dashboard/presentation/dashboard_page.dart';
+import 'package:librarian_app/src/features/dashboard/views/dashboard_page.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 
 class SplashPage extends StatefulWidget {
